+++ services/auth-service/src/main/java/com/example/auth_service/Repository/ProfileRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.example.auth_service.Entity.Profile;
 
-import java.util.Optional;
-
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
 }
