+++ services/auth-service/src/main/java/com/example/auth_service/Service/ProfileService.java
@@ -3,6 +3,6 @@ package com.example.auth_service.Service;
 import com.example.auth_service.DTO.ProfileDto;
 
 public interface ProfileService {
-    ProfileDto getProfile(Integer userId);
-    ProfileDto updateProfile(Integer userId, ProfileDto profileDto);
+    ProfileDto getProfile(Long userId);
+    ProfileDto updateProfile(Long userId, ProfileDto profileDto);
 }
